:recycle: move payload to separate file
