+++ libs/private/model/convs-mgr/micro-apps/cmi5/src/lib/functions/cmi5-parser-payload.interface.ts
@@ -1,0 +1,7 @@
+export interface CMI5ParserPayload
+{
+  bucketName: string;
+  courseId: string;
+  orgId: string;
+  fileName: string;
+}
